requirements.txt
Django==2.2.0
djangorestframework==3.11.0
